rename: change ga folder
